feat(marketplace): add SSR and checkout session
